Moved EditProfileFragment to views/users
