+++ BobsLittleFreeLibrary/app/src/main/java/com/example/bobslittlefreelibrary/views/users/PublicProfileViewActivity.java
@@ -1,6 +1,5 @@
 package com.example.bobslittlefreelibrary.views.users;
 
-import android.app.AppComponentFactory;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
@@ -9,15 +8,8 @@ import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.bobslittlefreelibrary.EditProfileFragment;
 import com.example.bobslittlefreelibrary.R;
 import com.example.bobslittlefreelibrary.models.User;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.FirebaseFirestore;
 
 public class PublicProfileViewActivity extends AppCompatActivity {
